routes: refact -> isolate useLocation() hook
