+++ packages/site-admin/src/router/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useLocation } from './use-location';
