+++ packages/site-admin/src/router/index.ts
@@ -1,3 +1,16 @@
+/**
+ * External dependencies
+ */
+import { createContext } from '@wordpress/element';
+/**
+ * Types
+ */
+import type { Config, Match } from './types';
+
+// Context instances
+export const RoutesContext = createContext< Match | null >( null );
+export const ConfigContext = createContext< Config >( { pathArg: 'p' } );
+
 // Components
 export * from './link';
 export * from './router';
